+++ Pinta.Core/Actions/EditActions.cs
@@ -280,7 +280,7 @@ namespace Pinta.Core
 				if (rect.Width == 0 || rect.Height == 0)
 					return;
 			
-				ImageSurface dest = new ImageSurface (Format.Argb32, rect.Width, rect.Height);
+				ImageSurface dest = CairoExtensions.CreateImageSurface (Format.Argb32, rect.Width, rect.Height);
 
 				using (Context g = new Context (dest)) {
 					g.SetSourceSurface (src, -rect.X, -rect.Y);
@@ -305,7 +305,8 @@ namespace Pinta.Core
 				var rect = doc.GetSelectedBounds (true);
 
 				// Copy it to a correctly sized surface 
-				using (var dest = new ImageSurface (Format.Argb32, rect.Width, rect.Height)) {
+				using (var dest = CairoExtensions.CreateImageSurface (Format.Argb32, rect.Width, rect.Height)) {
+
 					using (Context g = new Context (dest)) {
 						g.SetSourceSurface (src, -rect.X, -rect.Y);
 						g.Paint ();