+++ Pinta.Gui.Widgets/Widgets/Canvas/CanvasRenderer.cs
@@ -84,7 +84,7 @@ namespace Pinta.Gui.Widgets
 					if (scale_factor.Ratio == 1)
 						layer.Draw (g, layer.Surface, layer.Opacity, false);
 					else {
-						using (var scaled = new Cairo.ImageSurface (Cairo.Format.Argb32, dst.Width, dst.Height)) {
+						using (var scaled = CairoExtensions.CreateImageSurface (Cairo.Format.Argb32, dst.Width, dst.Height)) {
 							g.Save ();
 							// Have to undo the translate set above
 							g.Translate (offset.X, offset.Y);
@@ -107,12 +107,12 @@ namespace Pinta.Gui.Widgets
 			get {
 				// Create one if we don't have one
 				if (offset_layer == null)
-					offset_layer = new Layer (new Cairo.ImageSurface (Cairo.Format.ARGB32, source_size.Width, source_size.Height));
+					offset_layer = new Layer (CairoExtensions.CreateImageSurface (Cairo.Format.ARGB32, source_size.Width, source_size.Height));
 
 				// If we have the wrong size one, dispose it and create the correct size
 				if (offset_layer.Surface.Width != source_size.Width || offset_layer.Surface.Height != source_size.Height) {
 					(offset_layer.Surface as IDisposable).Dispose ();
-					offset_layer = new Layer (new Cairo.ImageSurface (Cairo.Format.ARGB32, source_size.Width, source_size.Height));
+					offset_layer = new Layer (CairoExtensions.CreateImageSurface (Cairo.Format.ARGB32, source_size.Width, source_size.Height));
 				}
 
 				return offset_layer;