+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Thanks to the following contributors who worked on this release:
 - @tdgroot
 - @Nikitayfh
 - @pikachuiscool2
+- @yaminb
 
 ### Added
 - Ported to GTK3 and .NET 5
@@ -63,6 +64,7 @@ Thanks to the following contributors who worked on this release:
 - Fixed issues where the zoom level was not maintained when resizing an image ([#1889673](https://bugs.launchpad.net/pinta/+bug/1889673)).
 - Fixed a bug where the initial corner of a rectangle shape could be cut off ([#1922470](https://bugs.launchpad.net/pinta/+bug/1922470)).
 - Fixed a bug where the text tool was not correctly clipped against the selection ([#1910511](https://bugs.launchpad.net/pinta/+bug/1910511)).
+- Improved handling of memory allocation failures for large images ([#776346](https://bugs.launchpad.net/pinta/+bug/776346)).
 
 ## [1.7](https://github.com/PintaProject/Pinta/releases/tag/1.7) - 2020/08/04